Criação dos testes unitarios do service

+++ src/main/java/com/mmelo/starwars/domain/Planet.java
@@ -31,5 +31,5 @@ public class Planet extends Auditable {
     private String terrain;
 
     @Column(name = "quantity_films")
-    private Long quantityFilms;
+    private int quantityFilms;
 }
